Display only events for which the end date is today or in future.
Note that the Sunday meetings aren't filtered like this because we want to show the whole calendar of the season at least for now. In future we could do some filtering though, e.g. show only the upcoming meetings and separately be able to display the meetings of the season.