chore: Update most of the packages, including gatsby
Note that unfortunately Gatsby seems to have somewhat of a dependency hell in its packages and it even recommends to use  parameter when running npm install. I don't want to do that so needed to set a couple of overrides.

Also recreated the package-lock.json from scratch which isn't ideal.

Set the used npm and node version (to the current LTS) to be explicit about them too since the newer Gatsby version raised the node version required (minimum 18 now so still far from the currently active LTS).